Move payment method helpers to composite-checkout dir
