+++ client/my-sites/checkout/checkout/composite-checkout/payment-method-helpers.js
@@ -11,17 +11,11 @@ import wp from 'lib/wp';
  * Internal dependencies
  */
 import {
-	WPCOMTransactionEndpoint,
-	WPCOMTransactionEndpointRequestPayload,
-	WPCOMTransactionEndpointResponse,
 	createTransactionEndpointRequestPayloadFromLineItems,
-	PayPalExpressEndpoint,
-	PayPalExpressEndpointRequestPayload,
-	PayPalExpressEndpointResponse,
 	createPayPalExpressEndpointRequestPayloadFromLineItems,
-} from './types';
+} from '../types';
 
-const debug = debugFactory( 'calypso:composite-checkout-payment-methods' );
+const debug = debugFactory( 'calypso:composite-checkout:payment-method-helpers' );
 
 export function useStoredCards( getStoredCards ) {
 	const [ state, dispatch ] = useReducer( storedCardsReducer, {
@@ -55,10 +49,7 @@ function storedCardsReducer( state, action ) {
 	}
 }
 
-export async function submitExistingCardPayment(
-	transactionData,
-	submit: WPCOMTransactionEndpoint
-): Promise< WPCOMTransactionEndpointResponse > {
+export async function submitExistingCardPayment( transactionData, submit ) {
 	debug( 'formatting existing card transaction', transactionData );
 	const formattedTransactionData = createTransactionEndpointRequestPayloadFromLineItems( {
 		...transactionData,
@@ -68,10 +59,7 @@ export async function submitExistingCardPayment(
 	return submit( formattedTransactionData );
 }
 
-export async function submitApplePayPayment(
-	transactionData,
-	submit: WPCOMTransactionEndpoint
-): Promise< WPCOMTransactionEndpointResponse > {
+export async function submitApplePayPayment( transactionData, submit ) {
 	debug( 'formatting apple-pay transaction', transactionData );
 	const formattedTransactionData = createTransactionEndpointRequestPayloadFromLineItems( {
 		...transactionData,
@@ -82,10 +70,7 @@ export async function submitApplePayPayment(
 	return submit( formattedTransactionData );
 }
 
-export async function makePayPalExpressRequest(
-	transactionData,
-	submit: PayPalExpressEndpoint
-): Promise< WPCOMTransactionEndpointResponse > {
+export async function makePayPalExpressRequest( transactionData, submit ) {
 	const formattedTransactionData = createPayPalExpressEndpointRequestPayloadFromLineItems( {
 		...transactionData,
 	} );
@@ -102,10 +87,7 @@ export async function fetchStripeConfiguration( requestArgs, wpcom ) {
 	return wpcom.stripeConfiguration( requestArgs );
 }
 
-export async function sendStripeTransaction(
-	transactionData,
-	submit: WPCOMTransactionEndpoint
-): Promise< WPCOMTransactionEndpointResponse > {
+export async function sendStripeTransaction( transactionData, submit ) {
 	const formattedTransactionData = createTransactionEndpointRequestPayloadFromLineItems( {
 		...transactionData,
 		paymentMethodToken: transactionData.paymentMethodToken.id,
@@ -116,10 +98,7 @@ export async function sendStripeTransaction(
 	return submit( formattedTransactionData );
 }
 
-export function submitCreditsTransaction(
-	transactionData,
-	submit: WPCOMTransactionEndpoint
-): Promise< WPCOMTransactionEndpointResponse > {
+export function submitCreditsTransaction( transactionData, submit ) {
 	debug( 'formatting full credits transaction', transactionData );
 	const formattedTransactionData = createTransactionEndpointRequestPayloadFromLineItems( {
 		...transactionData,
@@ -129,10 +108,7 @@ export function submitCreditsTransaction(
 	return submit( formattedTransactionData );
 }
 
-export function submitFreePurchaseTransaction(
-	transactionData,
-	submit: WPCOMTransactionEndpoint
-): Promise< WPCOMTransactionEndpointResponse > {
+export function submitFreePurchaseTransaction( transactionData, submit ) {
 	debug( 'formatting free transaction', transactionData );
 	const formattedTransactionData = createTransactionEndpointRequestPayloadFromLineItems( {
 		...transactionData,
@@ -142,7 +118,7 @@ export function submitFreePurchaseTransaction(
 	return submit( formattedTransactionData );
 }
 
-export function isPaymentMethodEnabled( method: string, allowedPaymentMethods: string[] ): boolean {
+export function isPaymentMethodEnabled( method, allowedPaymentMethods ) {
 	// By default, allow all payment methods
 	if ( ! allowedPaymentMethods?.length ) {
 		return true;
@@ -197,15 +173,11 @@ function WordPressLogo() {
 	);
 }
 
-export async function wpcomTransaction(
-	payload: WPCOMTransactionEndpointRequestPayload
-): Promise< WPCOMTransactionEndpointResponse > {
+export async function wpcomTransaction( payload ) {
 	return wp.undocumented().transactions( payload );
 }
 
-export async function wpcomPayPalExpress(
-	payload: PayPalExpressEndpointRequestPayload
-): Promise< PayPalExpressEndpointResponse > {
+export async function wpcomPayPalExpress( payload ) {
 	return wp.undocumented().paypalExpressUrl( payload );
 }
